+++ fili-core/src/main/java/com/yahoo/bard/webservice/data/config/metric/makers/ArithmeticMaker.java
@@ -92,7 +92,9 @@ public class ArithmeticMaker extends MetricMaker {
     protected LogicalMetric makeInner(String metricName, List<String> dependentMetrics) {
         // Get the ArithmeticPostAggregation operands from the dependent metrics
         List<PostAggregation> operands = dependentMetrics.stream()
-                .map(this::getNumericField)
+                .map(metrics::get)
+                .map(LogicalMetric::getMetricField)
+                .map(MetricMaker::getNumericField)
                 .collect(Collectors.toList());
 
         // Create the ArithmeticPostAggregation