+++ fili-core/src/main/java/com/yahoo/bard/webservice/data/config/metric/makers/SketchSetOperationMaker.java
@@ -47,8 +47,8 @@ public class SketchSetOperationMaker extends MetricMaker {
         TemplateDruidQuery mergedQuery = getMergedQuery(dependentMetrics);
 
         // Get the SketchSetOperationPostAggregation operands from the dependent metrics
-        PostAggregation operandOne = getSketchField(dependentMetrics.get(0));
-        PostAggregation operandTwo = getSketchField(dependentMetrics.get(1));
+        PostAggregation operandOne = MetricMaker.getSketchField(metrics.get(dependentMetrics.get(0)).getMetricField());
+        PostAggregation operandTwo = MetricMaker.getSketchField(metrics.get(dependentMetrics.get(1)).getMetricField());
 
         // Create the SketchSetOperationPostAggregation
         SketchSetOperationPostAggregation setPostAggregation;